Refactor test code to match new import style
