+++ tests/test_assistant.py
@@ -4,7 +4,7 @@ from alembic.operations import Operations
 from illuminate.exceptions import BasicManagerException
 from illuminate.manager import Assistant
 from illuminate.manager import Manager
-from tests.shared.unit import Test
+from tests.unit import Test
 
 
 class TestAssistantClass(Test):
