+++ tests/test_decorator_cli.py
@@ -3,7 +3,7 @@ from alembic.operations import Operations
 
 from illuminate.decorators import adapt
 from illuminate.manager import Manager
-from tests.shared.unit import Test
+from tests.unit import Test
 
 
 class DummyManager:
