+++ tests/test_exporter_sql.py
@@ -5,7 +5,7 @@ import pytest
 from illuminate.exceptions import BasicExporterException
 from illuminate.exporter import SQLExporter
 from illuminate.manager import Manager
-from tests.shared.unit import Test
+from tests.unit import Test
 
 
 class TestExporterSQL(Test):
